Added a threads version of the graph generator (slower)
The threads version is slower because I had it generate a thread for each graph attempt which can get into the millions very quickly so it was not a smart play. There is definitely a clever way to do it but I don't think it is worth the effort at this point. I believe I need to push through and work on different aspects of the project besides the graph generator.