Added Rust Graph Generator
This is my implementation of a graph generator that is programmed in rust. It also implements generating the graphs as an adjacency matrix instead of a 2D array of indeces in order to prevent isomorphism. This is still an extremely slow, inefficient, and impractical algorithm since it can't generate the number of permutations of a schedule within 20 minutes and I think the running time is some O(n^6) and I think higher but I'm not exactly sure about recursion and big O.